x/structs/types/message_substation_allocation_connect_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"structs/testutil/sample"
)

func TestMsgSubstationAllocationConnect_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSubstationAllocationConnect
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSubstationAllocationConnect{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSubstationAllocationConnect
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
